- Created Transform class - TransformComponent inherits from Transform
